refractored ipfs functions to one module
