--- src/Ipfs/ipfs.js
@@ -1,9 +1,0 @@
-import IPFS from "ipfs-api";
-
-const ipfs = new IPFS({
-  host: "ipfs.infura.io",
-  port: 5001,
-  protocol: "https",
-});
-
-export default ipfs;
